rename tooltipUtils to be consistent camelcase
