+++ src/modules/tutorial.js
@@ -4,7 +4,7 @@ import mapboxgl from "mapbox-gl";
 // es6 module import: https://github.com/js-cookie/js-cookie/issues/233
 window.Cookies = require("js-cookie");
 
-import { setTooltip, showTooltip } from "./tooltip_utils.js";
+import { setTooltip, showTooltip } from "./tooltipUtils.js";
 
 window.ATD_TUTORIAL = {};
 const BOOTSTRAP_SM_BREAKPOINT = 575;
